Feat: 뉴스 목록 및 Load More UI 완료, Template 파일로 서버 요청 영역 분리
