+++ src/App.js
@@ -1,12 +1,12 @@
 import { Reset } from 'styled-reset';
 import './App.css';
-import NewsList from './components/board/NewsList';
+import NewsTemplate from './components/board/NewsTemplate';
 
 function App() {
   return (
     <div className='content-wrapper'>
       <Reset />
-      <NewsList />
+      <NewsTemplate />
     </div>
   );
 }
