+++ src/styles/NewsTemplate.css
